+++ tests/series_only/head_test.py
@@ -2,13 +2,13 @@ from __future__ import annotations
 
 import pytest
 
-import narwhals.stable.v1 as nw_v1
+import narwhals as nw
 from tests.utils import ConstructorEager, assert_equal_data
 
 
 @pytest.mark.parametrize("n", [2, -1])
 def test_head_series(constructor_eager: ConstructorEager, n: int) -> None:
-    df = nw_v1.from_native(constructor_eager({"a": [1, 2, 3]}), eager_only=True)
+    df = nw.from_native(constructor_eager({"a": [1, 2, 3]}), eager_only=True)
     result = df.select(df["a"].head(n))
     expected = {"a": [1, 2]}
     assert_equal_data(result, expected)
